add cast to double

+++ bitherj/src/main/java/net/bither/bitherj/core/PeerManager.java
@@ -1001,7 +1001,7 @@ public class PeerManager {
         if (synchronizing && syncStartHeight > 0 && downloadingPeer != null && lastBlockHeight >=
                 syncStartHeight && lastBlockHeight <= downloadingPeer.getVersionLastBlockHeight()) {
             AbstractApp.notificationService.sendBroadcastProgressState((double) (lastBlockHeight
-                    - syncStartHeight) / (downloadingPeer.getVersionLastBlockHeight() -
+                    - syncStartHeight) / (double) (downloadingPeer.getVersionLastBlockHeight() -
                     syncStartHeight));
         } else {
             AbstractApp.notificationService.sendBroadcastProgressState(-1);
